requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
python-telegram-bot
pytrends
python-dotenv
schedule
pandas
openpyxl
requests
markdown
